Move `dotted_circle` check from Universal to Shaping profile
